requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
idna==3.4
ifaddr==0.2.0
influxdb==5.3.1
influxdb-client==1.36.1
kiwisolver==1.4.4
matplotlib==3.7.0
msgpack==1.0.5
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycodestyle==2.10.0
pycryptodome==3.17
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
pytz==2022.7.1
reactivex==4.0.4
requests==2.28.2
six==1.16.0
sonofflan==0.3.1
typing_extensions==4.5.0
urllib3==1.26.15
zeroconf==0.39.4
